Add files via upload
This project uses a  Random Forest classifier to distinguish between  oranges  and  grapefruits based on physical attributes (diameter, weight, and RGB values). Built with Streamlit, it provides a user-friendly interface for real-time predictions. Simply input the fruit's characteristics, and the app predicts whether it’s an orange or grapefruit. Ideal for quick and accurate citrus classification!